fix(notebooks): fix migration admin refresh and facade imports
