+++ products/notebooks/backend/facade/api.py
@@ -15,11 +15,12 @@ lives in ``facade.collab``.
 from typing import TYPE_CHECKING, Any
 from uuid import UUID
 
-from .. import logic
+from .. import logic, markdown_migration
 from ..models import Notebook
 from . import contracts
 
 if TYPE_CHECKING:
+    from posthog.models import User
     from posthog.rbac.user_access_control import UserAccessControl
 
 
@@ -78,6 +79,12 @@ def get_notebook_activity_summary(team_id: int, limit: int) -> contracts.Noteboo
             for row in recent
         ],
     )
+
+
+def get_markdown_notebook_migration_stats(
+    team_id: int | None = None,
+) -> contracts.MarkdownNotebookMigrationStats:
+    return markdown_migration.get_markdown_notebook_migration_stats(team_id)
 
 
 # --- Access control ---
@@ -150,6 +157,21 @@ def create_notebook(
         visibility=visibility,
     )
     return _to_notebook_data(notebook)
+
+
+def migrate_notebooks_to_markdown(
+    *,
+    user: "User",
+    team_id: int | None = None,
+    dry_run: bool = True,
+    max_previews: int = 5,
+) -> contracts.MarkdownNotebookMigrationResult:
+    return markdown_migration.migrate_notebooks_to_markdown(
+        user=user,
+        team_id=team_id,
+        dry_run=dry_run,
+        max_previews=max_previews,
+    )
 
 
 # --- Resource links (groups, accounts) ---
